backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.7.1
alembic==1.13.1
python-multipart==0.0.9
httpx==0.27.0
stripe==9.3.0
fastapi-mail==1.4.1
aiosmtplib==3.0.1
jinja2==3.1.4
